feat: Bump nextlove-sdk-generator to 1.8.5

+++ src/Objects/DeviceBattery.php
@@ -18,7 +18,7 @@ class DeviceBattery
 
     
     public function __construct(
-        public int $level,
+        public float $level,
     ) {
     }
